Añadidas validaciones, Mapper en el POM, Falta Añadir Email y Internacionalizar mensajes Error

+++ src/main/java/org/iesvdm/dao/ComercialDAOImpl.java
@@ -77,7 +77,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                 							  rs.getString("nombre"), 
                 							  rs.getString("apellido1"),
                 							  rs.getString("apellido2"), 
-                							  rs.getFloat("comision"))
+                							  rs.getBigDecimal("comision"))
                 						 	
         );
 		
@@ -94,7 +94,7 @@ public class ComercialDAOImpl implements ComercialDAO {
             						 						rs.getString("nombre"),
             						 						rs.getString("apellido1"),
             						 						rs.getString("apellido2"),
-            						 						rs.getFloat("comision"))
+            						 						rs.getBigDecimal("comision"))
 						,id
 								);
